requirements.txt
Django==1.11.9
psycopg2==2.6.2
gunicorn
Pillow==5.0.0
django-allauth==0.34.0
django-bootstrap4==0.0.6

